# ResetTrainOpenApi.yaml
openapi: 3.0.3
info:
  title: Reset Train - OpenAPI 3.0
  description: |-
    Servidor de Reset Train con especificaciones OpenAPI 3.0. 

  termsOfService: http://localhost:8080/gym/
  contact:
    email: resettrain@info.es
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8080/gym/
...
paths:
  /perfils:
    get:
      tags:
        - Perfil
      summary: Listar perfiles
      description: Lista todos los perfiles disponibles
      operationId: getPerfil
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Perfil'
        '500':
          description: Internal Server Error
          content:
            application/json:
              example:
                mensaje: "Ocurrió un error interno en el servidor."
  /perfils/{id}:
    get:
      tags:
        - Perfil
      summary: Buscar perfil por ID
      description: Devuelve un perfil
      operationId: getPerfilById
      parameters:
        - name: id
          in: path
          description: ID de la perfil a devolver
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                $ref: '#/components/schemas/Perfil'
        '400':
          description: Bad Request
          content:
            application/json:
              example:
                mensaje: "El id debe ser un número."
        '404':
          description: Perfil no encontrada
          content:
            application/json:
              example:
                mensaje: "La rutina solicitada no fue encontrada."
        '500':
          description: Error interno del servidor
          content:
            application/json:
              example:
                mensaje: "Ocurrió un error interno en el servidor."
  /perfiles:
    get:
      tags:
        - Perfil
      summary: Listar perfiles
      description: Lista todos los perfiles disponibles
      operationId: getPerfiles
      parameters:
        - name: obesidad
          in: query
          description: Muestra todos los perfiles con obesidad
          required: false
          explode: true
          schema:
            type: boolean
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Perfil'
        '500':
          description: Internal Server Error
          content:
            application/json:
              example:
                mensaje: "Ocurrió un error interno en el servidor."
    post:
      tags:
        - Perfil
      summary: Añade una nuevo perfil
      description: Añade una nuevo perfil a la bbdd
      operationId: addPerfil
      requestBody:
        description: Crea una nuevo perfil
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perfil'
          application/xml:
            schema:
              $ref: '#/components/schemas/Perfil'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Perfil'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                $ref: '#/components/schemas/Perfil'
        '400':
          description: Bad Request
          content:
            application/json:
              example:
                mensaje: "Argumento no válido."
        '500':
          description: Internal Server Error
          content:
            application/json:
              example:
                mensaje: "Ocurrió un error interno en el servidor."
  /perfil/{id}:
    delete:
      tags:
        - Perfil
      summary: Borra un perfil por ID
      description: borra un perfil por ID
      operationId: deletePerfil
      parameters:
        - name: id
          in: path
          description: Perfil id a borrar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Sin contenido(exito al borrar)
        '404':
          description: Not Found
          content:
            application/json:
              example:
                mensaje: "El ID no existe."
        '500':
          description: Internal Server Error
          content:
            application/json:
              example:
                mensaje: "Ocurrió un error interno en el servidor."
    put:
      tags:
        - Perfil
      summary: Actualiza un perfil por ID
      description: Actualiza por ID
      operationId: modifyPerfil
      parameters:
        - name: id
          in: path
          description: ID del perfil a actualizar
          required: true
          schema:
            type: integer
            format: int64
        - name: fechaMedicion
          in: query
          description: fecha de la medición
          schema:
            type: string
            format: date-time
        - name: obesidad
          in: query
          description: Si la persona tiene obesidad
          schema:
            type: boolean
        - name: ritmoCardiaco
          in: query
          description: ritmo cardiaco
          schema:
            type: integer
            format: int32
        - name: imc
          in: query
          description: Imc del usuario
          schema:
            type: integer
            format: int32
        - name: peso
          in: query
          description: Peso del usuario
          schema:
            type: integer
            format: int32
        - name: medidas
          in: query
          description: medidas del usuario
          schema:
            type: integer
            format: int32
      requestBody:
        description: Actualiza un perfil existente en la bbdd
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerfilBody'
          application/xml:
            schema:
              $ref: '#/components/schemas/PerfilBody'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PerfilBody'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                $ref: '#/components/schemas/Perfil'
        '400':
          description: Bad Request
          content:
            application/json:
              example:
                mensaje: "Argumento no válido."
        '404':
          description: Not Found
          content:
            application/json:
              example:
                mensaje: "El ID no existe."
        '500':
          description: Internal Server Error
          content:
            application/json:
              example:
                mensaje: "Ocurrió un error interno en el servidor."
...
components:
  schemas:
    Perfil:
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        ritmoCardiaco:
          type: integer
          format: int32
          example: 85
        peso:
          type: integer
          format: int32
          example: 80
        imc:
          type: integer
          format: int32
          example: 25
        obesidad:
          type: boolean
          example: true
        fechaMedicion:
          type: string
          format: date-time
          example: 2023-05-06
        medidas:
          type: string
          example: 97-80-85
      xml:
        name: perfil
...
    PerfilBody:
      type: object
      properties:
        ritmoCardiaco:
          type: integer
          format: int32
          example: 85
        peso:
          type: integer
          format: int32
          example: 80
        imc:
          type: integer
          format: int32
          example: 25
        obesidad:
          type: boolean
          example: true
        fechaMedicion:
          type: string
          format: date-time
          example: 2023-05-06
        medidas:
          type: string
          example: 97-80-85
      xml:
        name: PerfilBody
...